requirements.txt
behave
allure-behave
pandas
numpy
mysql-connector-python
google-cloud-bigquery
pyarrow
so